fix: PlanMetrics serialization + LONG type para consecutive_failures
- bronze/ingest: usa spark.table() para S3 upload (DF limpo sem plan metrics)
- agent_pre: state table com LONG ao inves de INT (Python int -> Spark LONG)

+++ pipeline/notebooks/agent_pre.py
@@ -43,7 +43,7 @@ spark.sql(f"""
         run_at                STRING,
         last_bronze_hash      STRING,
         status                STRING,
-        consecutive_failures  INT,
+        consecutive_failures  LONG,
         delta_versions        STRING
     )
     USING DELTA
